requirements.txt
pandas
openpyxl
Pillow
pyinstaller
